Rename extension: added support to change output extension
When using 'classic' toolchain (with csc and mono compilers),
the extension of executables was set to '.exe' (even on Linux).
With dotnet core cli, it is not possible to change output extension easily.
So, to allow keep compatibility with other tools which expects an '.exe'
extension, this commit enables the client to specify the desired extension
through a new compilation flag. Example: '-D outputExtension=exe'.